Main project files moved to src folder
