+++ WorkoutTracker.Api.sln
@@ -1,22 +1,39 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35514.174 d17.12
+VisualStudioVersion = 17.12.35514.174
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WorkoutTracker.Api", "WorkoutTracker.Api.csproj", "{9F60E3AC-75F8-4379-A186-1BCB1C99A70C}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WorkoutTracker.Api", "src\WorkoutTracker.Api\WorkoutTracker.Api.csproj", "{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9F60E3AC-75F8-4379-A186-1BCB1C99A70C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9F60E3AC-75F8-4379-A186-1BCB1C99A70C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9F60E3AC-75F8-4379-A186-1BCB1C99A70C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9F60E3AC-75F8-4379-A186-1BCB1C99A70C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Debug|x64.Build.0 = Debug|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Debug|x86.Build.0 = Debug|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Release|Any CPU.Build.0 = Release|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Release|x64.ActiveCfg = Release|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Release|x64.Build.0 = Release|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Release|x86.ActiveCfg = Release|Any CPU
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{688E73D6-AF5F-4EC6-82D2-9A6FDBEA8015} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
