+++ src/WorkoutTracker.Api/DTOs/Auth/RegisterDto.cs
