+++ src/WorkoutTracker.Api/DTOs/Exercise/ExerciseReadDto.cs
