+++ src/WorkoutTracker.Api/DTOs/Exercise/ExerciseUpdateDto.cs
