+++ src/WorkoutTracker.Api/DTOs/TrainingSession/Exercise/TrainingExerciseReadDto.cs
