+++ src/WorkoutTracker.Api/DTOs/TrainingSession/TrainingSession/TrainingSessionUpdateDto.cs
