+++ src/WorkoutTracker.Api/Exceptions/EntityAlreadyExistsException.cs
