+++ src/WorkoutTracker.Api/Exceptions/EntityNotFoundException.cs
