+++ src/WorkoutTracker.Api/Models/ApplicationUser.cs
