+++ src/WorkoutTracker.Api/Models/Exercise.cs
