+++ src/WorkoutTracker.Api/Models/TrainingExercise.cs
