+++ src/WorkoutTracker.Api/Models/TrainingSession.cs
