+++ src/WorkoutTracker.Api/Services/Auth/IAuthService.cs
