+++ src/WorkoutTracker.Api/WorkoutTracker.Api.csproj
@@ -24,8 +24,4 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Services\" />
-  </ItemGroup>
-
 </Project>
